fix: move interfaces to the interfaces folder
