+++ client/src/components/RootComponents/rightSideBar/MyMatches.tsx
@@ -1,10 +1,10 @@
-import { Typography } from '@mui/material';
 import { ReactElement } from 'react';
+
+import { Typography } from '@mui/material';
+
 import { StyledButton, StyledMyMatches } from '../../index';
 
-interface MyMatchesProps {
-  stadium: string;
-}
+import { MyMatchesProps } from '../../../interfaces';
 
 const MyMatches = ({ stadium }: MyMatchesProps): ReactElement => {
   return (
