+++ client/src/interfaces/rootLayout.ts
@@ -3,3 +3,20 @@ export interface UserCardProps {
   userId: number;
   role: string;
 }
+
+export interface SideBarLinkProps {
+  text: string;
+  icon: React.ReactNode;
+}
+
+export interface MyMatchesProps {
+  stadium: string;
+}
+
+export interface RightSideBarTitleProps {
+  title: string;
+}
+
+export interface TeamNameProps {
+  text: string;
+}
